requirements.txt
requests
Pillow
bs4
pyowm
aiogram
fuzzywuzzy